Refactor logic out of const.py to keep it leaner
