+++ custom_components/ef_powerocean_tcpmodbus/const.py
@@ -2,9 +2,6 @@
 
 from __future__ import annotations
 
-from collections.abc import Iterable, Sequence
-from dataclasses import dataclass
-from enum import StrEnum
 from typing import Final
 
 from homeassistant.const import (
@@ -16,6 +13,21 @@ from homeassistant.const import (
     UnitOfPower,
     UnitOfRatio,
     UnitOfTemperature,
+)
+
+from .models import (
+    BinarySensorDef,
+    CoordinatorStatus,
+    EnergySensorDef,
+    GridMode,
+    InverterModel,
+    NumberWritableDef,
+    OperatingMode,
+    RegisterBlock,
+    RegisterDef,
+    RegisterType,
+    SensorDef,
+    plan_blocks,
 )
 
 DOMAIN: Final = "ef_powerocean_tcpmodbus"
@@ -43,220 +55,13 @@ MAX_BATTERY_DISCHARGED_POWER: Final = 3300
 MAX_BATTERY_COUNT: Final = 12
 MAX_FAULT_EVENTS: Final = 20
 
-MAX_REGISTERS_PER_READ: Final = 125
-# Reading a few unused registers is cheaper than a second round trip, so registers
-# closer together than this share one request.
-MAX_REGISTER_GAP: Final = 48
-
 SLEEP_TIME_AFTER_RECONNECT_S: Final = 1
 SLEEP_TIME_AFTER_BATTERY_CHECK_FAILED_S: Final = 15
 ENERGY_RESOLUTION_KWH: Final = 0.01
 STORAGE_VERSION: Final = 1
 STATE_SAVE_DELAY_S: Final = 30
 
-
-class InverterModel(StrEnum):
-    POWEROCEAN_SINGLE_PHASE = "powerocean_single_phase"
-    POWEROCEAN_THREE_PHASE = "powerocean_three_phase"
-    POWEROCEAN_PLUS = "powerocean_plus"
-    POWEROCEAN_DC_FIT = "powerocean_dc_fit"
-    OCEAN_2 = "ocean_2"
-
-    @property
-    def startup_voltage(self) -> int:
-        return {
-            # The startup voltage is used to filter out phantom string power when the PV input is not actually producing power.
-            # The values are based on the datasheets of each model. However, the single phase does not have a dedicated startup voltage specification
-            # so this value is deducted from the MPPT operating range.
-            # https://enterprise-service-eu-cdn.ecoflow.com/enterprise/content/2024-03-27-1485da5d-eae4-4a38-830a-4e340517d968.pdf
-            self.POWEROCEAN_SINGLE_PHASE: 90,
-            # https://enterprise-service-eu-cdn.ecoflow.com/enterprise/documentation/1772090325968/EcoFlow%20PowerOcean%20(Three-phase)_Datasheet_EN.pdf
-            self.POWEROCEAN_THREE_PHASE: 160,
-            # https://enterprise-service-eu-cdn.ecoflow.com/enterprise/documentation/1754035729875/PowerOcean%20Plus%20(three-phase)_Brochure_20241223_EN.pdf
-            self.POWEROCEAN_PLUS: 160,
-            # https://enterprise-service-eu-cdn.ecoflow.com/enterprise/documentation/1735192805714/EcoFlow%20PowerOcean%20DC%20Fit_Datasheet_EN_20241225.pdf
-            self.POWEROCEAN_DC_FIT: 90,
-            # https://enterprise-service-eu-cdn.ecoflow.com/enterprise/documentation/1779447439219/OCEAN%202%20Three-Phase_Datasheet_EN_260522.pdf
-            self.OCEAN_2: 120,
-        }[self]
-
-    @property
-    def display_name(self) -> str:
-        return {
-            self.POWEROCEAN_SINGLE_PHASE: "PowerOcean Single Phase",
-            self.POWEROCEAN_THREE_PHASE: "PowerOcean Three Phase",
-            self.POWEROCEAN_PLUS: "PowerOcean Plus",
-            self.POWEROCEAN_DC_FIT: "PowerOcean DC Fit",
-            self.OCEAN_2: "Ocean 2",
-        }[self]
-
-    @classmethod
-    def from_product_info(
-        cls, product_number: int | None, product_category: int | None
-    ) -> InverterModel | None:
-        """Map the device's product registers to a model, or None if unknown.
-
-        We are not sure of the product number for the remaining models. Feel free
-        to contribute this if you own such a model.
-        """
-        if product_number == 1:
-            return (
-                cls.POWEROCEAN_SINGLE_PHASE
-                if product_category == 2
-                else cls.POWEROCEAN_THREE_PHASE
-            )
-        if product_number == 2:
-            return cls.POWEROCEAN_SINGLE_PHASE
-        if product_number == 3:
-            return cls.POWEROCEAN_PLUS
-        return None
-
-
 DEFAULT_INVERTER_MODEL: Final = InverterModel.POWEROCEAN_THREE_PHASE
-
-
-class CoordinatorStatus(StrEnum):
-    SUCCESS = "success"
-    READ_FAILED = "read_failed"
-    RECONNECT_FAILED = "reconnect_failed"
-    PROCESSING_FAILED = "processing_failed"
-
-
-class OperatingMode(StrEnum):
-    STANDBY = "standby"
-    SELF_CONSUMPTION = "self_consumption"
-    AI = "ai"
-    UNKNOWN = "unknown"
-
-
-class GridMode(StrEnum):
-    GRID = "grid"
-    ISLANDED = "islanded"
-
-
-class RegisterType(StrEnum):
-    """Word layout of a register. Multi-word values are stored low word first."""
-
-    UINT16 = "uint16"
-    UINT32 = "uint32"
-    FLOAT32 = "float32"
-    SERIAL = "serial"
-
-
-REGISTER_SIZES: Final = {
-    RegisterType.UINT16: 1,
-    RegisterType.UINT32: 2,
-    RegisterType.FLOAT32: 2,
-    # 16 ASCII bytes.
-    RegisterType.SERIAL: 8,
-}
-
-
-@dataclass(frozen=True)
-class RegisterDef:
-    key: str
-    address: int
-    data_type: RegisterType = RegisterType.FLOAT32
-
-    @property
-    def size(self) -> int:
-        """Return how many 16-bit words this register occupies."""
-        return REGISTER_SIZES[self.data_type]
-
-    @property
-    def end(self) -> int:
-        """Return the address just past this register."""
-        return self.address + self.size
-
-
-@dataclass(frozen=True)
-class RegisterBlock:
-    """Registers that are fetched with a single Modbus request."""
-
-    registers: tuple[RegisterDef, ...]
-
-    def __post_init__(self) -> None:
-        if self.count > MAX_REGISTERS_PER_READ:
-            raise ValueError(
-                f"Block at {self.start} spans {self.count} registers, "
-                f"more than the {MAX_REGISTERS_PER_READ} a Modbus read allows."
-            )
-
-    @property
-    def start(self) -> int:
-        return min(register.address for register in self.registers)
-
-    @property
-    def count(self) -> int:
-        return max(register.end for register in self.registers) - self.start
-
-    def index_of(self, register: RegisterDef) -> int:
-        """Return the register's offset within this block's response."""
-        return register.address - self.start
-
-    def registers_for(self, raw: Sequence[int], register: RegisterDef) -> list[int]:
-        """Return the raw words of *register* within this block's response."""
-        index = self.index_of(register)
-        return list(raw[index : index + register.size])
-
-
-def plan_blocks(registers: Iterable[RegisterDef]) -> tuple[RegisterBlock, ...]:
-    """Group registers into the fewest Modbus reads.
-
-    A new read starts when the next register is too far away to be worth reading
-    through, or when the block would outgrow a single Modbus response.
-    """
-    blocks: list[RegisterBlock] = []
-    current: list[RegisterDef] = []
-
-    for register in sorted(registers, key=lambda register: register.address):
-        if current:
-            gap = register.address - max(mapped.end for mapped in current)
-            span = register.end - current[0].address
-            if gap > MAX_REGISTER_GAP or span > MAX_REGISTERS_PER_READ:
-                blocks.append(RegisterBlock(tuple(current)))
-                current = []
-        current.append(register)
-
-    if current:
-        blocks.append(RegisterBlock(tuple(current)))
-    return tuple(blocks)
-
-
-@dataclass(frozen=True)
-class SensorDef:
-    key: str
-    name: str | None = None
-    unit: str | None = None
-    device_class: str | None = None
-    state_class: str | None = None
-    entity_category: EntityCategory | None = None
-    icon: str | None = None
-    options: tuple[str, ...] | None = None
-
-
-@dataclass(frozen=True)
-class EnergySensorDef:
-    key: str
-    name: str | None = None
-    unit: str = UnitOfEnergy.KILO_WATT_HOUR
-    is_calculated: bool = False
-    resets_daily: bool = False
-    max_power: int | None = None
-    # The _total value of the energy counter, used to validate daily resets and prevent invalid spikes.
-    total_source: str | None = None
-    device_class: str = "energy"
-    state_class: str = "total_increasing"
-    entity_category: EntityCategory | None = None
-    icon: str | None = None
-
-
-@dataclass(frozen=True)
-class BinarySensorDef:
-    key: str
-    name: str | None = None
-    device_class: str | None = None
-    entity_category: EntityCategory | None = None
 
 
 PRODUCT_CATEGORY: Final = RegisterDef("product_category", 40002, RegisterType.UINT16)
@@ -718,30 +523,13 @@ BINARY_SENSOR_MAP: list[BinarySensorDef] = [
 ]
 
 
-@dataclass(frozen=True)
-class NumberWritableDef:
-    key: str  # Unique key for the number entity (e.g., "min_soc_limit_control")
-    read_key: str  # The original key from MODBUS_REGISTERS used for reading (e.g., "min_soc_limit")
-    name: str  # Display name for Home Assistant UI
-    min_value: float  # Slider minimum value
-    max_value: float  # Slider maximum value
-    step: float  # Step size (1.0 for integers, 0.1 for floats)
-    unit: str | None = None  # Unit of measurement
-    device_class: str | None = None  # Device class type
-    icon: str | None = None  # Custom icon for the slider
-
-    @property
-    def register(self) -> int:
-        """Write back to the same address the value is read from."""
-        return REGISTERS_BY_KEY[self.read_key].address
-
-
 # Map of all modbus registers available for writing operations.
 WRITABLE_NUMBERS_MAP: list[NumberWritableDef] = [
     NumberWritableDef(
         key="min_soc_limit_control",
         read_key="min_soc_limit",
         name="Minimum SOC Limit Control",
+        register=REGISTERS_BY_KEY["min_soc_limit"].address,
         min_value=0.0,
         max_value=100.0,
         step=1.0,
@@ -752,6 +540,7 @@ WRITABLE_NUMBERS_MAP: list[NumberWritableDef] = [
         key="device_led_brightness_control",
         read_key="device_led_brightness",
         name="LED Brightness Control",
+        register=REGISTERS_BY_KEY["device_led_brightness"].address,
         min_value=0.0,
         max_value=100.0,
         step=10.0,
