+++ tests/test_telemetry.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import unittest
 
 import pytest
-from ef_powerocean_tcpmodbus.const import RegisterType
+from ef_powerocean_tcpmodbus.models import RegisterType
 from ef_powerocean_tcpmodbus.telemetry import (
     TelemetryData,
     calculate_derived_values,
